# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Food Facts Taxonomy Editor API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/{taxonomy_name}/{branch}/project": {
      "get": {
        "summary": "Get Project Info",
        "description": "Get information about a Taxonomy Editor project",
        "operationId": "get_project_info__taxonomy_name___branch__project_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/set-project-status": {
      "get": {
        "summary": "Set Project Status",
        "description": "Set the status of a Taxonomy Editor project",
        "operationId": "set_project_status__taxonomy_name___branch__set_project_status_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ProjectStatus" },
                { "type": "null" }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/nodes": {
      "get": {
        "summary": "Find All Nodes",
        "description": "Get all nodes within taxonomy",
        "operationId": "find_all_nodes__taxonomy_name___branch__nodes_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Node",
        "description": "Deleting given node from a taxonomy",
        "operationId": "delete_node__taxonomy_name___branch__nodes_delete",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/rootentries": {
      "get": {
        "summary": "Find All Root Nodes",
        "description": "Get all root nodes within taxonomy",
        "operationId": "find_all_root_nodes__taxonomy_name___branch__rootentries_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/entry/{entry}": {
      "get": {
        "summary": "Find One Entry",
        "description": "Get entry corresponding to id within taxonomy",
        "operationId": "find_one_entry__taxonomy_name___branch__entry__entry__get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "entry",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entry" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Edit Entry",
        "description": "Editing an entry in a taxonomy.\nNew key-value pairs can be added, old key-value pairs can be updated.\nURL will be of format '/entry/<id>'",
        "operationId": "edit_entry__taxonomy_name___branch__entry__entry__post",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "entry",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entry" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/entry/{entry}/parents": {
      "get": {
        "summary": "Find One Entry Parents",
        "description": "Get parents for a entry corresponding to id within taxonomy",
        "operationId": "find_one_entry_parents__taxonomy_name___branch__entry__entry__parents_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "entry",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entry" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/entry/{entry}/children": {
      "get": {
        "summary": "Find One Entry Children",
        "description": "Get children for a entry corresponding to id within taxonomy",
        "operationId": "find_one_entry_children__taxonomy_name___branch__entry__entry__children_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "entry",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entry" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Edit Entry Children",
        "description": "Editing an entry's children in a taxonomy.\nNew children can be added, old children can be removed.\nURL will be of format '/entry/<id>/children'",
        "operationId": "edit_entry_children__taxonomy_name___branch__entry__entry__children_post",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "entry",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Entry" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/synonym/{synonym}": {
      "get": {
        "summary": "Find One Synonym",
        "description": "Get synonym corresponding to id within taxonomy",
        "operationId": "find_one_synonym__taxonomy_name___branch__synonym__synonym__get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "synonym",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Synonym" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Edit Synonyms",
        "description": "Editing a synonym in a taxonomy.\nNew key-value pairs can be added, old key-value pairs can be updated.\nURL will be of format '/synonym/<id>'",
        "operationId": "edit_synonyms__taxonomy_name___branch__synonym__synonym__post",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          },
          {
            "name": "synonym",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Synonym" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/{taxonomy_name}/{branch}/synonym": {
      "get": {
        "summary": "Find All Synonyms",
        "description": "Get all synonyms within taxonomy",
        "operationId": "find_all_synonyms__taxonomy_name___branch__synonym_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/{taxonomy_name}/{branch}/stopword": {
      "get": {
        "summary": "Find All Stopwords",
        "description": "Get all stopwords within taxonomy",
        "operationId": "find_all_stopwords__taxonomy_name___branch__stopword_get",
        "parameters": [
          {
            "name": "branch",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Branch" }
          },
          {
            "name": "taxonomy_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Taxonomy Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Project": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "status": {
            "allOf": [{ "$ref": "#/components/schemas/ProjectStatus" }],
            "default": "LOADING"
          },
          "taxonomyName": { "type": "string", "title": "Taxonomyname" },
          "branchName": { "type": "string", "title": "Branchname" },
          "description": { "type": "string", "title": "Description" },
          "ownerName": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Ownername"
          },
          "isFromGithub": { "type": "boolean", "title": "Isfromgithub" },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "errorsCount": {
            "type": "integer",
            "title": "Errorscount",
            "default": 0
          },
          "githubCheckoutCommitSha": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Githubcheckoutcommitsha"
          },
          "githubFileLatestSha": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Githubfilelatestsha"
          },
          "githubPrUrl": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Githubprurl"
          }
        },
        "type": "object",
        "required": [
          "id",
          "taxonomyName",
          "branchName",
          "description",
          "isFromGithub",
          "createdAt"
        ],
        "title": "Project"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": ["OPEN", "EXPORTED", "LOADING", "FAILED"],
        "title": "ProjectStatus"
      },
...
    }
  }
}